Huge feature updates

+++ app/Http/Controllers/JudgingRuleController.php
@@ -4,36 +4,20 @@ namespace App\Http\Controllers;
 
 use App\JudgingRule;
 use Illuminate\Http\Request;
+use App\Competition;
 
-class JudgingRuleController extends Controller {
+class JudgingRuleController extends Controller
+{
 
     /**
      * Create a new judging rule
      * @param Request $request
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function create(Request $request) {
-        $request->validate([
-            'name' => 'required|max:255',
-            'description' => 'required',
-            'score' => 'required|integer|min:1',
-            'weight' => 'required|integer|min:1',
-            'group' => 'required'
-        ]);
-
-        $rule = new JudgingRule();
-        $rule->fill($request->all())->save();
-        return redirect('/portal/rubric')->with('success', "Rule created.");
-    }
-
-    /**
-     * Update a judging rule
-     * @param Request $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function update(Request $request, $id) {
-        $rule = JudgingRule::find($id);
-        if($rule){
+    public function create(Request $request, $competition_id)
+    {
+        $competition = Competition::find($competition_id);
+        if ($competition && $competition->status !== 'archived') {
             $request->validate([
                 'name' => 'required|max:255',
                 'description' => 'required',
@@ -42,10 +26,42 @@ class JudgingRuleController extends Controller {
                 'group' => 'required'
             ]);
 
-            $rule->fill($request->all())->save();
-            return redirect('/portal/rubric')->with('success', "Rule updated.");
+            $rule = new JudgingRule();
+            $rule->fill($request->all());
+            $rule->competition_id = $competition->id;
+            $rule->save();
+            return redirect('/portal/rubric/' . $competition->id)->with('success', "Rule created.");
         } else {
-            return back()->with('error', "Rule not found.");
+            return response("404", 404);
+        }
+    }
+
+    /**
+     * Update a judging rule
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function update(Request $request, $competition_id, $id)
+    {
+        $competition = Competition::find($competition_id);
+        if ($competition && $competition->status !== 'archived') {
+            $rule = JudgingRule::where('id', $id)->where('competition_id', $competition_id)->first();
+            if ($rule) {
+                $request->validate([
+                    'name' => 'required|max:255',
+                    'description' => 'required',
+                    'score' => 'required|integer|min:1',
+                    'weight' => 'required|integer|min:1',
+                    'group' => 'required'
+                ]);
+
+                $rule->fill($request->all())->save();
+                return redirect('/portal/rubric/' . $competition->id)->with('success', "Rule updated.");
+            } else {
+                return back()->with('error', "Rule not found.");
+            }
+        } else {
+            return response("404", 404);
         }
     }
 
@@ -54,13 +70,19 @@ class JudgingRuleController extends Controller {
      * @param Request $request
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function delete(Request $request){
-        $rule = JudgingRule::find($request->id);
-        if($rule){
-            $rule->delete();
-            return redirect('/portal/rubric')->with('success', "Rule deleted.");
+    public function delete(Request $request, $competition_id)
+    {
+        $competition = Competition::find($competition_id);
+        if ($competition && $competition->status !== 'archived') {
+            $rule = JudgingRule::where('id', $request->id)->where('competition_id', $competition_id)->first();
+            if ($rule) {
+                $rule->delete();
+                return redirect('/portal/rubric')->with('success', "Rule deleted.");
+            } else {
+                return back()->with('error', 'Cannot find rule.');
+            }
         } else {
-            return back()->with('error', 'Cannot find rule.');
+            return response("404", 404);
         }
     }
 }
